benchmarks/face_detection/requirements_gpu.txt
# =============================================================================
# Requirements GPU pour RunPod - Face Detection Benchmark
# =============================================================================
# Installation: pip install -r requirements_gpu.txt
# Optimisé pour RunPod avec CUDA 12.1+
# =============================================================================

# -----------------------------------------------------------------------------
# Core
# -----------------------------------------------------------------------------
numpy>=1.21.0
opencv-python>=4.8.0
scipy>=1.9.0
tqdm>=4.65.0
Pillow>=9.0.0

# -----------------------------------------------------------------------------
# PyTorch avec CUDA (installé séparément pour compatibilité CUDA)
# -----------------------------------------------------------------------------
# Note: PyTorch est installé via le script setup avec:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0

# -----------------------------------------------------------------------------
# ONNX Runtime GPU
# -----------------------------------------------------------------------------
onnxruntime-gpu>=1.15.0

# -----------------------------------------------------------------------------
# Frameworks de détection
# -----------------------------------------------------------------------------
mediapipe>=0.10.0
mtcnn>=0.1.1
facenet-pytorch>=2.5.0
insightface>=0.7.3
ultralytics>=8.1.0

# -----------------------------------------------------------------------------
# Téléchargement
# -----------------------------------------------------------------------------
gdown>=4.7.0
requests>=2.28.0
huggingface_hub>=0.19.0

# -----------------------------------------------------------------------------
# Optionnel (commenter si problèmes de compilation)
# -----------------------------------------------------------------------------
# dlib>=19.24.0
